requirements.txt
fastapi[standard]
uvicorn
bcrypt
pyjwt
sqlalchemy
boto3
python-multipart
python-dotenv

docx
requests
python-docx

pandas